+++ src/main/java/de/telran/onlineshopgarden/repository/OrderItemRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
+public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
 
-    boolean existsByProductId(Integer productId);
+    boolean existsByProductId(Long productId);
 }
